requirements.txt
beautifulsoup4==4.10.0
nltk==3.6.6
odfpy==1.4.1
pdfplumber==0.5.28
slate3k==0.5.3
tabulate==0.8.9
pytesseract==0.3.10
pdf2image==1.17.0